Fix compilation issues on Solaris 10 with GCC 4.5
git-svn-id: svn://svn.code.sf.net/p/libkqueue/code/trunk@518 fb4e3144-bc1c-4b72-a658-5bcd248dd7f7

+++ src/common/debug.h
@@ -32,7 +32,7 @@ extern char *KQUEUE_DEBUG_IDENT;
 # include <sys/syscall.h>
 # define THREAD_ID ((pid_t)  syscall(__NR_gettid))
 #elif defined(__sun)
-# define THREAD_ID (pthread_self())
+# define THREAD_ID ((int) pthread_self())
 #elif defined(_WIN32)
 # define THREAD_ID (int)(GetCurrentThreadId())
 #else
